Move class to another package as its used by ServerResponse
