+++ core/src/main/scala/sttp/tapir/server/interceptor/CustomInterceptors.scala
@@ -1,5 +1,6 @@
 package sttp.tapir.server.interceptor
 
+import sttp.tapir.server.ValuedEndpointOutput
 import sttp.tapir.server.interceptor.content.UnsupportedMediaTypeInterceptor
 import sttp.tapir.server.interceptor.decodefailure.{DecodeFailureHandler, DecodeFailureInterceptor, DefaultDecodeFailureHandler}
 import sttp.tapir.server.interceptor.exception.{DefaultExceptionHandler, ExceptionHandler, ExceptionInterceptor}
